backend/requirements.txt
annotated-types==0.7.0
ansible==11.1.0
ansible-core==2.18.1
ansible-runner==2.4.0
anyio==4.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
Jinja2==3.1.4
lockfile==0.12.2
MarkupSafe==3.0.2
packaging==24.2
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
pycurl==7.45.4
pydantic==2.10.4
pydantic_core==2.27.2
python-daemon==3.1.2
python-dotenv==1.0.1
python-terraform==0.10.1
PyYAML==6.0.2
requests==2.32.3
resolvelib==1.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
SQLAlchemy>=1.4
asyncpg==0.30.0
alembic==1.15.2
python-jose==3.4.0
passlib[bcrypt]
bcrypt==4.3.0
jose==1.0.0
jwt==1.3.1
email-validator
pydantic[email]
cachetools